Create knapsack problem using greedy techniques
Output :

Enter the no. of objects:- 7

Enter the wts and profits of each object:-
2 10
3 5
5 15
7 7
1 6
4 18
1 3

Enter the capacity of knapsack:- 15

The result vector is:- 1.000000        1.000000        1.000000        1.000000
    1.000000        0.666667        0.000000

Maximum profit is:- 55.333332
1
2
3
4
5
6
7
8
9
10
11
12
13
14
15
16
17
Enter the no. of objects:- 7
 
Enter the wts and profits of each object:-
2 10
3 5
5 15
7 7
1 6
4 18
1 3
 
Enter the capacity of knapsack:- 15
 
The result vector is:- 1.000000        1.000000        1.000000        1.000000
    1.000000        0.666667        0.000000
 
Maximum profit is:- 55.333332